+++ tests/tests/partition_isolation.rs
@@ -57,6 +57,7 @@ async fn write_agent_shard(
         Arc::clone(&store) as Arc<dyn ailake_store::Store>,
         policy(Some("agent_id".into()), Some(agent.into())),
         table.clone(),
+        2,
     )
     .await
     .unwrap();
@@ -284,6 +285,7 @@ async fn score_fn_is_invoked_during_search() {
         Arc::clone(&store) as Arc<dyn ailake_store::Store>,
         policy(None, None),
         table.clone(),
+        2,
     )
     .await
     .unwrap();
@@ -343,6 +345,7 @@ async fn score_fn_constant_zero_returns_rows() {
         Arc::clone(&store) as Arc<dyn ailake_store::Store>,
         policy(None, None),
         table.clone(),
+        2,
     )
     .await
     .unwrap();